Crate event

A library for event threading

Reexports

pub use generic_event::GenericEvent;

Modules

events

generic_event
Trait for generic events

Structs

Events
An event loop iterator

IdleArgs
Idle arguments, such as expected idle time in seconds.

MaxFps
The maximum number of frames per second

RenderArgs
Render arguments

UpdateArgs
Update arguments, such as delta time in seconds

Ups
The number of updates per second

Enums

Event
Adds render and update events to input events

Traits

FocusEvent
When window gets or looses focus

IdleEvent
When background tasks should be performed

MouseCursorEvent
The position of the mouse cursor

MouseRelativeEvent
The relative movement of mouse cursor

MouseScrollEvent
The scroll of the mouse wheel

PressEvent
The press of a button

ReleaseEvent
The release of a button

RenderEvent
When the next frame should be rendered

ResizeEvent
When the window is resized

TextEvent
When receiving text from user, such as typing a character

UpdateEvent
When the application state should be updated

Functions

events
Creates event iterator from window.