Module event::events

Structs

Events
An event loop iterator

IdleArgs
Idle arguments, such as expected idle time in seconds.

MaxFps
The maximum number of frames per second

PollEvent
Polls event from window.

RenderArgs
Render arguments

ShouldClose
Whether window should close or not.

Size
The size of the window.

SwapBuffers
Tells window to swap buffers.

UpdateArgs
Update arguments, such as delta time in seconds

Ups
The number of updates per second

Constants

DEFAULT_MAX_FPS
The default maximum frames per second.

DEFAULT_UPS
The default updates per second.

Traits

EventMap
Methods required to map from consumed event to emitted event.